Updates scripts to use `npx grunt`.
`npx` will try running the package as installed in `node_modules/.bin`,
(if available), falling back to a globally-installed package (if
available), and finally downloading, running, and deleting the package
if it's not available locally.

In this application, as long as the user ran `npm install`, there
should be a copy of the grunt CLI script in `node_modules/.bin`.